fix: Move Predictions and Labels types to `types` module
